chore: axios interceptors modifications done.

+++ client/src/utils/interceptors.ts
@@ -1,6 +1,4 @@
 import axios from "axios";
-import { useDispatch } from "react-redux";
-import { setLoggedIn } from "src/redux/slices/auth";
 import { Notify } from "src/components";
 const jwtInterceptor = () => {
   axios.interceptors.request.use((request: any) => {
@@ -17,15 +15,18 @@ const responseInterceptor = (callback: () => void) => {
   axios.interceptors.response.use(
     (response) => response,
     (error) => {
+      console.log("==>> ", error);
+      if (!error.response) {
+        // Show a generic error Toast (e.g. Server error, please retry)
+        return Promise.reject(error);
+      }
       if (error.response.status === 401) {
         console.log(error);
         localStorage.clear();
-        const dispatch = useDispatch();
-        dispatch(setLoggedIn({ loggedIn: false }));
-        window.location.href = "/login";
-        callback();
         Notify(`${error.response.data}`, "error");
+        callback();
       }
+      return Promise.reject(error);
     }
   );
 };
